+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This project adheres to [Semantic Versioning](https://semver.org).
 ### Improvements
 - Move the cursor on the first character of the currently selected autocompletion suggestion, to highlight it
 
+### Fixes
+- Move to the last available character when deleting a line that is longer that the previous one, with `d`
+
 ## [0.3.3] - 2022/05/01
 
 ### Features